corrigido

+++ Lab3/ex4.c
@@ -6,14 +6,14 @@ int main(void)
 	float max = FLT_MAX;
 	float min = FLT_MIN;
 
-	double dmax = LDBL_MAX;
-	double dmin = LDBL_MIN;
+	double dmax = DBL_MAX;
+	double dmin = DBL_MIN;
 
-	printf("%-20f %-5e %-5E\n", max, max, max);
-	printf("%-20f %-5e %-5E\n", min, min, min);
+	printf("%5f %-5e %-5E\n", max, max, max);
+	printf("%5f %-5e %-5E\n", min, min, min);
 
-	printf("%20f %-5e %-5E\n", dmax, dmax, dmax);
-	printf("%-20f %-5e %-5E\n", dmin, dmin, dmin);
+	printf("%5f %-5e %-5E\n", dmax, dmax, dmax);
+	printf("%5f %-5e %-5E\n", dmin, dmin, dmin);
 
 	return 0;
